requirements.txt
streamlit
langchain
langchain_chroma
langchain_core
langchain_groq
langchain_huggingface
langchain_text_splitters
langchain_community
python-dotenv
faiss-cpu
google-generativeai
requests
